refactor: clean up background
